fix test -> pin fiona and  linux path
