+++ envs/hydromt-test.yml
@@ -12,6 +12,7 @@ dependencies:
   - coverage == 6.4 # tests
   - dask
   - entrypoints
+  - fiona == 1.8.22 # tmp pin because of incorrect CRS parsing
   - gdal>=3.1
   - geopandas>=0.10
   - netcdf4
